requirements.txt
# Mental Health Detection notebook dependencies
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
xgboost
shap

# App UI and report
streamlit
reportlab
Pillow
tensorflow
python-decouple

# Optional: SMS OTP (forgot password)
twilio
